+++ core/src/main/java/com/cobber/fta/token/FloatToken.java
@@ -40,4 +40,9 @@ public class FloatToken extends Token {
 	public int charactersUsed() {
 		return 1;
 	}
+
+	@Override
+	public boolean isComplete() {
+		return true;
+	}
 }